requirements.txt
streamlit>=1.28.0
mysql-connector-python>=8.0.33
pandas>=2.0.0
plotly>=5.17.0
matplotlib>=3.7.0